requirements.txt
digitalai-release-sdk == 24.1.0
